Only install dependencies if they are not already installed
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@6934 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ buildtools/scripts/make_dist.sh
@@ -10,8 +10,10 @@ else
   source ~/.simgrid_build.conf
 fi
 
-echo "get Linux dependencies"
-sudo aptitude install -y libtool automake1.10 autoconf libgcj10-dev gcc g++ bash flex flexml doxygen bibtex bibtool iconv bibtex2html addr2line valgrind transfig
+if [ ! -e /usr/bin/libtool ] || [ ! -e /usr/bin/automake ] || [ ! -e /usr/bin/flex ] ; then
+  echo "get Linux dependencies"
+  sudo aptitude install -y libtool automake1.10 autoconf libgcj10-dev gcc g++ bash flex flexml doxygen bibtex bibtool iconv bibtex2html addr2line valgrind transfig
+fi
 
 make_dist
